# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"get":{"operationId":"AuthController_access_token","summary":"log the user in with the intra and set the cookie","parameters":[],"responses":{"200":{"description":"the route responsible of fetching the authenticated user data from the intra API"}},"tags":["auth"]}},"/auth/log-out":{"delete":{"operationId":"AuthController_logout","summary":"log out and clear cookie","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","summary":"get user profile","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/users/upload_avatar":{"post":{"operationId":"UsersController_uploadFile","summary":"Change a user's avatar","parameters":[],"responses":{"200":{"description":"The uploaded avatar Details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvatarDto"}}}}},"tags":["users"]}},"/users/update_username":{"post":{"operationId":"UsersController_updateUsername","summary":"Change a user's username","parameters":[],"responses":{"201":{"description":""}},"tags":["users"]}},"/users/add_friend":{"post":{"operationId":"UsersController_AddFriend","summary":"Add a friend to a user","parameters":[],"responses":{"201":{"description":""}},"tags":["users"]}},"/users/friends_list":{"get":{"operationId":"UsersController_friendsList","summary":"get friends list","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/users/blocked_list":{"get":{"operationId":"UsersController_blockedFriend","summary":"get friends list","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getUserData","summary":"Get user data by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/authentication_page":{"get":{"operationId":"AppController_getAuthPage","summary":"Redirect to the authorization page of the intra Api","parameters":[],"responses":{"200":{"description":""}},"tags":["app"]}},"/two-factor-authentication/generate":{"post":{"operationId":"TwoFactorAuthenticationController_register","summary":"Generate the Qrcode","parameters":[],"responses":{"201":{"description":""}},"tags":["two-factor-authentication"]}},"/two-factor-authentication/turn-on":{"post":{"operationId":"TwoFactorAuthenticationController_turnOnTwoFacAuth","summary":"Enable 2FA","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFacAuthCodeDto"}}}},"responses":{"200":{"description":""}},"tags":["two-factor-authentication"]}},"/two-factor-authentication/authenticate":{"post":{"operationId":"TwoFactorAuthenticationController_authenticate","summary":"Validate the 2FA code and set the cookie","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwoFacAuthCodeDto"}}}},"responses":{"200":{"description":""}},"tags":["two-factor-authentication"]}}},"info":{"title":"pong pong API","description":"The pong API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AvatarDto":{"type":"object","properties":{}},"User":{"type":"object","properties":{}},"TwoFacAuthCodeDto":{"type":"object","properties":{}}}}}